api 테스트 예시용 추가, api 테스트를 위한 파일 추가 와 수정
