+++ src/index.ts
@@ -1,20 +1,3 @@
-import express, { NextFunction, Request, Response } from "express";
-import bodyParser from "body-parser";
-import resume from "./resume";
-import errorHandler from "./middlewares/errorhandler";
+import app from "./app";
 
-const app = express();
-
-app.use(bodyParser.json());
-app.use(errorHandler);
-
-app.use("/resume", resume);
-
-app.post("/", (request: Request, response: Response) => {
-  console.log(request.body);
-  response.send(request.body.data);
-});
-
-app.listen(3000, () => {
-  console.log("server started on port 3000");
-});
+app.listen(3000);
